+++ desktop/php/Abeille-TopButtons.php
@@ -11,7 +11,7 @@
 		$tools = array(
 			'health'    => array( 'bouton'=>'bt_healthAbeille',		'icon'=>'fa-medkit',        'text'=>'{{Santé}}' ),
 			'netList'   => array( 'bouton'=>'bt_network',			'icon'=>'fa-sitemap',       'text'=>'{{Réseau}}' ),
-			'net'       => array( 'bouton'=>'bt_networkMap',        'icon'=>'fa-map',           'text'=>'{{Placement réseau (BETA)}}' ),
+			'net'       => array( 'bouton'=>'bt_networkMap',        'icon'=>'fa-map',           'text'=>'{{Placement réseau}}' ),
 			'compat'    => array( 'bouton'=>'bt_supportedEqList',	'icon'=>'fa-align-left',    'text'=>'{{Compatibilité}}' ),
 			'ota'       => array( 'bouton'=>'bt_Ota',				'icon'=>'fa-paperclip',     'text'=>'{{Mises-à-jour OTA}}' ),
 			'support'   => array( 'bouton'=>'bt_maintenancePage',	'icon'=>'fa-medkit',        'text'=>'{{Maintenance}}' ),